requirements.txt
anyio==3.6.1
beautifulsoup4==4.11.1
certifi==2022.6.15.2
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.83.0
greenlet==1.1.3
h11==0.13.0
idna==3.4
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
numpy==1.23.3
pandas==1.4.4
protobuf==3.20.1
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
